pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "typer_configparse"
authors = [{name = "Ian A. Neal", email = "wopr@wopr.tech"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version", "description"]
dependencies = [
    "typer >=0.7.0,<1",
    "toml >=0.10.2,<1",
    "PyYAML >=6.0,<7",
    "pydantic >=1.10.2,<2",
]

[project.urls]
Home = "https://github.com/nealian/typer-configparse"
Documentation = "https://typer-configparse.readthedocs.io/"
